Improve namespace resolving based on an instance
In short, this scans for the usage of any `instance` functions. If it finds them, it tries to set up
the namespace resolving based on that `instance` HTML element: it might have an
`xpath-default-namespace` attribute that defines how XPaths on that instance should behave.

Note: some tests are still failing